Adaptations response modeling

To model adaptation induced by each training session.

adaptation_fn(data, k1, tau1, vars)

Arguments

data
A data frame object that contains at least training loads, performances and time between two consecutive sessions.

k1
The gain term for adaptation occurrences.

tau1
The time constant of the adaptations exponential decrease.

vars
A list that contains input (i.e. session training loads) and time (i.e. the time between two consecutive inputs) numerical vectors.

Value

a vector of numerical values

Author

Frank Imbach frankimbach@gmail.com

Examples

adaptation_fn(data = example_data, k1 = 0.5, tau1 = 40, vars = list("input" = example_data$training_load, "time" = example_data$rest))

#>   [1]     0.0000   436.9891   847.4735  1325.7830  1638.2715  2108.6875
#>   [7]  2632.9107  2942.1283  3245.7214  3718.5838  4386.3219  4783.6153
#>  [13]  5037.3184  4838.0054  5160.0802  5344.4310  5824.3073  6165.6107
#>  [19]  6412.8730  6496.8012  6873.1911  7291.4810  7352.9421  7669.6044
#>  [25]  7971.8371  8049.7024  7994.6488  8110.2540  8253.8369  7967.1658
#>  [31]  8112.4683  7991.9720  7873.3815  7908.4243  8026.8205  8237.8934
#>  [37]  8103.5139  8557.7900  8623.3401  9066.4539  8507.5943  8480.5609
#>  [43]  8425.3176  8571.1532  8745.9439  8743.7249  8628.0679  8513.8281
#>  [49]  9081.0723  9151.1178  9186.1308  9173.8669  8471.8967  8692.2609
#>  [55]  8656.5245  8975.8977  9066.2877  8753.6398  8758.6056  8751.2883
#>  [61]  8512.0549  8742.5241  8977.6988  9001.3333  9296.4029  9538.0315
#>  [67]  9519.4578  9859.9748  9464.7031  9254.5413  9491.2139  9457.5390
#>  [73]  9261.5423  9504.6618  9336.3440  9277.1907  9527.6853  9677.9656
#>  [79]  9168.0809  9283.6650  8602.9197  8487.5584  8898.9710  9199.5971
#>  [85]  9268.3568  9657.4657  9334.5436  8970.8061  9174.8014  9441.7406
#>  [91]  9278.1581  9991.6258  9928.3181  9929.4778 10461.1396 10228.6850
#>  [97] 10470.0319 10515.1307  9830.7008  9289.6512  9635.2320  9397.3373
#> [103]  9165.3162  8939.0238  8718.3185  8503.0624  8293.1211  8088.3632
#> [109]  7888.6608  7693.8890  7503.9263  7318.6537  7137.9554  6961.7187
#> [115]  6789.8333  6622.1917  6458.6892  6299.2236  6143.6952  5992.0068
#> [121]  5844.0636  5699.7732  5559.0453  5421.7920  5287.9275  5157.3681
#> [127]  5030.0322  4905.8403  4784.7146  4666.5796  4551.3613  4438.9878
#> [133]  4329.3888  4222.4958  4118.2420  4016.5623  3917.3930  3820.6722
#> [139]  3726.3395  3634.3358  3544.6038  3457.0872  3371.7314  3288.4831
#> [145]  3207.2901  3128.1019  3050.8687  2975.5425  2902.0761  2830.4236