Fatigue response modeling.

To calculate the variable fatigue component for each training input.

fatigue_fn(data, k3, tau2, tau3, vars)

Arguments

data
A data frame object that contains at least training loads, performances and time between two consecutive sessions.

k3
The gain term for fatigue occurrences.

tau2
The time constant of fatigue exponential decrease.

tau3
The time constant of the fatigue remanence exponential decrease.

vars
A list that contains input (i.e. session training loads) and time (i.e. the time between two consecutive inputs) numerical vectors.

Value

a vector of numerical values

Author

Frank Imbach frankimbach@gmail.com

Examples

fatigue_fn(data = example_data, k3=5, tau2 = 20, tau3 = 5, vars = list("input" = example_data$training_load, "time" = example_data$rest))

#>   [1]         0         0   2295455   7112344  12277810  17143554  24003958
#>   [8]  29391563  33938035  41090922  54094239  68405349  79073255  73159772
#>  [15]  75888819  76417305  83916964  91898135  98612867 100221029 108130296
#>  [22] 119856903 122330821 129848483 138616904 140843008 136586760 134549456
#>  [29] 134152173 120878489 117621957 108811301  99410966  94572066  92736620
#>  [36]  94461532  89797277  97019137  99235878 110239009  99744566  95486546
#>  [43]  91831788  91880670  94108677  93732162  90480592  86688974  95782250
#>  [50]  99946228 102303743 102303602  87097116  87015263  84737460  89316609
#>  [57]  92094300  86732927  84889679  83140954  77414535  79683201  84004097
#>  [64]  86068709  92952729 100245703 102925699 112689168 107912301 101607241
#>  [71] 104205774 102700761  97148926  99331087  95737314  93271081  96515218
#>  [78] 100067639  91336229  90975639  77324759  72155810  75388538  81829077
#>  [85]  86246176  95917862  93768104  86306410  87452890  92204983  90393323
#>  [92] 104614547 109670427 112456603 126417379 126213436 133292957 136006055
#>  [99] 119516311 103841600 104853061  99739317  94874973  90247866  85846426
#> [106]  81659646  77677058  73888704  70285109  66857264  63596597  60494954
#> [113]  57544580  54738098  52068489  49529079  47113518  44815764  42630074
#> [120]  40550980  38573286  36692044  34902552  33200335  31581135  30040905
#> [127]  28575793  27182135  25856447  24595413  23395880  22254850  21169468
#> [134]  20137021  19154927  18220730  17332094  16486798  15682728  14917872
#> [141]  14190319  13498249  12839931  12213721  11618050  11051431  10512447
#> [148]   9999749   9512055   9048147