Variation in the fatiguing effects.

To make the fatigue impulse response varying over time in accordance with system input.

k2i_fn(data, k3, tau3, vars)

Arguments

data
A data frame object that contains at least training loads, performances and time between two consecutive sessions.

k3
The gain term for fatigue occurrences.

tau3
The time constant of the fatigue remanence exponential decrease.

vars
A list that contains input (i.e. session training loads) and time (i.e. the time between two consecutive inputs) numerical vectors.

Value

a vector of numerical values

Author

Frank Imbach frankimbach@gmail.com

Examples

k2i_fn(data = example_data, k3=5, tau3 = 5, vars = list("input" = example_data$training_load, "time" = example_data$rest))

#>   [1] 0.000000e+00 2.793898e+03 5.065873e+03 8.129759e+03 5.937923e+03
#>   [6] 7.088770e+03 9.327116e+03 8.641737e+03 8.920415e+03 9.519929e+03
#>  [11] 1.726538e+04 1.872405e+04 1.671913e+04 7.702637e+03 9.847023e+03
#>  [16] 9.364862e+03 1.284710e+04 1.440935e+04 1.502602e+04 1.172741e+04
#>  [21] 1.460279e+04 1.749840e+04 1.424477e+04 1.537687e+04 1.736886e+04
#>  [26] 1.492027e+04 1.198945e+04 1.135859e+04 1.169102e+04 7.879274e+03
#>  [31] 9.075767e+03 7.687531e+03 6.843815e+03 7.481314e+03 8.565411e+03
#>  [36] 1.026481e+04 8.145639e+03 1.263417e+04 1.237936e+04 1.626498e+04
#>  [41] 8.507697e+03 8.236222e+03 8.026351e+03 9.375639e+03 1.072388e+04
#>  [46] 1.001402e+04 8.725185e+03 7.963016e+03 1.365823e+04 1.281877e+04
#>  [51] 1.176298e+04 1.116443e+04 5.694272e+03 8.304275e+03 8.277887e+03
#>  [56] 1.129616e+04 1.146071e+04 8.047024e+03 8.399675e+03 8.460212e+03
#>  [61] 6.864034e+03 9.319669e+03 1.144263e+04 1.090951e+04 1.308717e+04
#>  [66] 1.484034e+04 1.332312e+04 1.597548e+04 9.901910e+03 8.244163e+03
#>  [71] 1.066154e+04 9.839400e+03 8.214504e+03 1.056757e+04 9.014105e+03
#>  [76] 8.800116e+03 1.104681e+04 1.236521e+04 7.442280e+03 8.988232e+03
#>  [81] 5.008098e+03 5.642935e+03 9.825998e+03 1.211646e+04 1.156707e+04
#>  [86] 1.474600e+04 9.907304e+03 7.185001e+03 9.430114e+03 1.177403e+04
#>  [91] 9.788281e+03 1.760601e+04 1.443998e+04 1.317968e+04 1.859532e+04
#>  [96] 1.359818e+04 1.474676e+04 1.408609e+04 7.746809e+03 5.417655e+03
#> [101] 9.262001e+03 7.583085e+03 6.208505e+03 5.083094e+03 4.161685e+03
#> [106] 3.407300e+03 2.789661e+03 2.283981e+03 1.869966e+03 1.530999e+03
#> [111] 1.253476e+03 1.026259e+03 8.402298e+02 6.879220e+02 5.632229e+02
#> [116] 4.611279e+02 3.775396e+02 3.091033e+02 2.530724e+02 2.071981e+02
#> [121] 1.696395e+02 1.388891e+02 1.137127e+02 9.310012e+01 7.622393e+01
#> [126] 6.240687e+01 5.109443e+01 4.183258e+01 3.424962e+01 2.804122e+01
#> [131] 2.295821e+01 1.879659e+01 1.538935e+01 1.259973e+01 1.031579e+01
#> [136] 8.445852e+00 6.914879e+00 5.661424e+00 4.635182e+00 3.794966e+00
#> [141] 3.107055e+00 2.543842e+00 2.082721e+00 1.705188e+00 1.396090e+00
#> [146] 1.143022e+00 9.358271e-01 7.661904e-01 6.273036e-01 5.135928e-01