Variable dose-response modeling

The function models the performance according to Busso Variable dose-response model \(Busso, 2003\) and based on previously defined parameters.

perf_model(data, P0, k1, k3, tau1, tau2, tau3, vars, target)

Arguments

data
A data frame object that contains at least training loads, performances and time between two consecutive sessions.

P0
The basic level of performance.

k1
The gain term for adaptations response.

k3
The gain term for fatigue response.

tau1
The time constant of the adaptations exponential decrease.

tau2
The time constant of fatigue exponential decrease.

tau3
The time constant of the fatigue remanence exponential decrease.

vars
A list that contains input (i.e. session training loads) and time (i.e. the time between two consecutive inputs) numerical vectors.

target
A character vector that indicates the performances column name.

Value

a vector of numerical values

Author

Frank Imbach frankimbach@gmail.com

Examples

perf_model(data = example_data, P0 = 10, k1 = 0.1, k3 = 0.01, tau1 = 40, tau2 = 20, tau3 = 5, vars = list("input" = example_data$training_load, "time" = example_data$rest),
target = "perf")

#>   [1]      10.00000      97.39783   -4945.37419  -16689.75128  -18290.38064
#>   [6]  -30797.47162  -52498.77725  -57990.81471  -68489.21685  -84045.02044
#>  [11] -171067.81932 -204236.64417 -198150.84571  -97797.00321 -129479.71373
#>  [16] -128414.27408 -183036.43514 -215381.85985 -234046.89643 -191308.77660
#>  [21] -245161.51412 -305453.49767 -258405.99492 -287999.41781 -334178.04825
#>  [26] -296150.71181 -242702.18892 -231886.28070 -239501.76099 -161678.12041
#>  [31] -180132.27713 -150167.71079 -129263.48394 -135841.07126 -151856.31469
#>  [36] -180858.98539 -144579.73776 -222282.93838 -221759.54492 -297875.36569
#>  [41] -159628.36901 -151939.71876 -145949.93886 -167465.69724 -191556.65477
#>  [46] -180572.90359 -157650.28226 -142558.73391 -240720.28172 -232068.77805
#>  [51] -217771.67479 -208928.08940 -106564.88398 -146011.05882 -144117.93457
#>  [56] -195072.86587 -201210.91482 -143158.09507 -147162.88560 -146689.54912
#>  [61] -117817.28734 -156430.99212 -193194.33424 -187588.29686 -228913.08511
#>  [66] -265814.65091 -244564.06687 -300291.39205 -191720.52188 -158860.30098
#>  [71] -201886.38419 -187071.78960 -155578.93080 -196419.85874 -168379.50693
#>  [76] -163005.96224 -202523.69237 -228779.50833 -139483.92938 -163752.12798
#>  [81]  -89960.83955  -94482.89662 -156156.99391 -196960.94155 -194532.43877
#>  [86] -255384.25351 -177499.82118 -129077.85168 -165029.27845 -207138.62540
#>  [91] -174917.62389 -316878.42400 -269429.46290 -252015.33317 -363005.97497
#>  [96] -273693.29896 -302938.72982 -295108.98692 -163878.91468 -110604.88284
#> [101]       0.00000       0.00000       0.00000       0.00000       0.00000
#> [106]       0.00000       0.00000       0.00000       0.00000       0.00000
#> [111]       0.00000       0.00000       0.00000       0.00000       0.00000
#> [116]       0.00000       0.00000       0.00000       0.00000       0.00000
#> [121]       0.00000       0.00000       0.00000       0.00000       0.00000
#> [126]       0.00000       0.00000       0.00000       0.00000       0.00000
#> [131]       0.00000       0.00000       0.00000       0.00000       0.00000
#> [136]       0.00000       0.00000       0.00000       0.00000       0.00000
#> [141]       0.00000       0.00000       0.00000       0.00000       0.00000
#> [146]       0.00000       0.00000       0.00000       0.00000       0.00000