Extract observations

This function extracts known performances from a data frame object.

real_perf(data, target)

Arguments

data
A data frame object that contains at least training loads, performances and time between two consecutive sessions.

target
A character vector that indicates the performances column name.

Value

a vector of numerical values

Author

Frank Imbach frankimbach@gmail.com

Examples

real_perf(data = example_data, target = "perf")

#>   [1] 10.00000 10.20202 10.40404 10.60606 10.80808 11.01010 11.21212 11.41414
#>   [9] 11.61616 11.81818 12.02020 12.22222 12.42424 12.62626 12.82828 13.03030
#>  [17] 13.23232 13.43434 13.63636 13.83838 14.04040 14.24242 14.44444 14.64646
#>  [25] 14.84848 15.05051 15.25253 15.45455 15.65657 15.85859 16.06061 16.26263
#>  [33] 16.46465 16.66667 16.86869 17.07071 17.27273 17.47475 17.67677 17.87879
#>  [41] 18.08081 18.28283 18.48485 18.68687 18.88889 19.09091 19.29293 19.49495
#>  [49] 19.69697 19.89899 20.10101 20.30303 20.50505 20.70707 20.90909 21.11111
#>  [57] 21.31313 21.51515 21.71717 21.91919 22.12121 22.32323 22.52525 22.72727
#>  [65] 22.92929 23.13131 23.33333 23.53535 23.73737 23.93939 24.14141 24.34343
#>  [73] 24.54545 24.74747 24.94949 25.15152 25.35354 25.55556 25.75758 25.95960
#>  [81] 26.16162 26.36364 26.56566 26.76768 26.96970 27.17172 27.37374 27.57576
#>  [89] 27.77778 27.97980 28.18182 28.38384 28.58586 28.78788 28.98990 29.19192
#>  [97] 29.39394 29.59596 29.79798 30.00000  0.00000  0.00000  0.00000  0.00000
#> [105]  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000
#> [113]  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000
#> [121]  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000
#> [129]  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000
#> [137]  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000
#> [145]  0.00000  0.00000  0.00000  0.00000  0.00000  0.00000