Dose-response modeling

This function trains a dose-response model (Busso 2003) with or without validation procedure.

sysmod(
  data,
  vars,
  target,
  date_ID,
  specify = NULL,
  validation.method = "none",
  specs = NULL
)

Arguments

data
A data frame object that contains at least training loads, performances and time between two consecutive sessions.

vars
A list that contains input (i.e. session training loads, numeric) and time (i.e. the time between two consecutive inputs) vectors.

target
A character that indicates the performances column name.

date_ID
A character that indicates the date time object name.

specify
Default is "NULL". Alternatively, a list that contains "theta_init" numeric vector that contains initial values for P0, k1, k3, tau1, tau2, tau3 parameters, a numeric vector for lower and upper bounds named "lower" and "upper" respectively, a character defining the method used for parameter optimization "optim.method" (see details). Hence, PO denotes the initial level of performance. The first performance can be extracted through the function init_perf.

validation.method
Default is "none". Alternatively, simple data splitting or cross-validation can be specified (see details).

specs
Default is "NULL". If a validation method is specified, a list that contains splitting arguments has to be specified (see details)

Value

A list describing the model output and its performances.

Details

validation.method can be used to train and evaluate model within a cross-validation procedure. Methods available are "none", "simple" and "TS-CV" for no validation, a simple data split into training / testing sets and time-series cross-validation respectively. specs lets define data splitting specifications. For the "simple" method, specify decimals as the proportion of the data to be used for model training in "initialWindow", the proportion of data used for model evaluation in "horizon" and logical term for the "fixedWindow" within a list. For the "TS-CV" method, specify numeric values for "initialWindow", "horizon" and logical term for the "fixedWindow" within a list. Each of the optimization algorithm with constraints and used by optimx can be used in "optim.method".

Note

Model performances (RMSE, MAE and R squared) are calculated on data used for validation = c("simple", "TS-CV").

References

Busso T (2003). “Variable dose-response relationship between exercise training and performance.” Med. Sci. Sports Exerc., 35, 1188--1195.

Author

Frank Imbach frankimbach@gmail.com

Examples